+++ p5-analysis/tests/Library.test.ts
@@ -33,6 +33,9 @@ describe('Library.find', () => {
         importPath: 'https://unpkg.com/p5.vector-arguments@1.0.0'
       })
     ).toBeInstanceOf(Library);
+    expect(
+      Library.find({ importPath: 'https://unpkg.com/undefined-library-name' })
+    ).toBeNull();
   });
 });
 
@@ -50,5 +53,17 @@ describe('Library.inferLibrariesFromScripts', () => {
     expect(inferLibraries(`ml5.poseNet.js`)).toEqual(['ml5.js']);
     expect(inferLibraries(`p5.Pulse.js`)).toEqual(['p5.sound']);
     expect(inferLibraries(`p5.Speech.js`)).toEqual(['p5.speech']);
+  });
+
+  test('infers libraries from comment directives', () => {
+    const libs = Library.inferFromScripts([
+      `${testfilesPath}/library-config-comments.js`
+    ]);
+    expect(libs[0]).toBe(Library.find({ name: 'p5.sound' }));
+    const importPaths = libs.map(lib => lib.importPath);
+    expect(importPaths.slice(1)).toEqual([
+      'https://unpkg.com/@mediapipe/pose',
+      'https://cdn.jsdelivr.net/npm/@tensorflow-models/pose-detection'
+    ]);
   });
 });